+++ dotnet-source/ks.fiks.io.arkivsystem.sample/Handlers/JournalpostHentHandler.cs
@@ -1,0 +1,39 @@
+using System.Collections.Generic;
+using System.IO;
+using System.Reflection;
+using System.Xml.Schema;
+using System.Xml.Serialization;
+using KS.Fiks.IO.Arkiv.Client.Models.Innsyn.Hent;
+using KS.Fiks.IO.Client.Models;
+using Serilog;
+
+namespace ks.fiks.io.arkivsystem.sample.Handlers
+{
+    public class JournalpostHentHandler : BaseHandler
+    {
+        private static readonly ILogger Log = Serilog.Log.ForContext(MethodBase.GetCurrentMethod()?.DeclaringType);
+
+        public static JournalpostHent GetPayload(MottattMeldingArgs mottatt, XmlSchemaSet xmlSchemaSet,
+            out bool xmlValidationErrorOccured, out List<List<string>> validationResult)
+        {
+            JournalpostHent journalpostHent = null;
+            if (mottatt.Melding.HasPayload)
+            {
+                var text = GetPayloadAsString(mottatt, xmlSchemaSet, out xmlValidationErrorOccured,
+                    out validationResult);
+                Log.Information("Parsing journalpostHent: {Xml}", text);
+                if (string.IsNullOrEmpty(text))
+                {
+                    Log.Error("Tom sok? Text: {Sok}", text);
+                }
+
+                using var textReader = (TextReader)new StringReader(text);
+                journalpostHent = (JournalpostHent) new XmlSerializer(typeof(JournalpostHent)).Deserialize(textReader);
+            }
+
+            xmlValidationErrorOccured = false;
+            validationResult = null;
+            return journalpostHent;
+        }
+    }
+}
